+++ CHANGELOG.md
@@ -1,6 +1,8 @@
+# Changelog
+
 ## Unreleased
 
-No unreleased changes.
+- Removed `kp get` and `kp set` commands, basically we removed the global config. Config can only be done via `kanpai.json` or the `kanpai` field in `package.json`.
 
 ## 0.10.1